==Changelog== - Added better semantic tags - Added titles to buttons with only icons to better convey its purpose
==Testing==
- Changed themes
- Started timer
- added tasks
- hovered over icons to see title description

==Version Compatibility==

==Detail Description==
- Tried to better follow the semantics adhering of HTML by adding in
semantic tags. Replaced a div in the navigation with a nav tag. Added
in a main tag to wrap the content and separated each section with a
section tag instead of a div.
- added titles to the following buttons, focus, reset, settings, help,
add task, toggle task compeletion, focus task, and delete task. The
content of the title is the function of each button.